requirements.txt
requests~=2.31.0
discord.py~=2.3.2
